[enhancement] retry patch with interpret mode when screen off
